64-bit
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@143 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxSDL/Helper.cpp
@@ -52,7 +52,7 @@ RTSEMEVENT    g_EventSemXPCOMQueueThread = NULL;
  */
 DECLCALLBACK(int) xpcomEventThread(RTTHREAD thread, void *pvUser)
 {
-    int eqFD = (int)pvUser;
+    int eqFD = (intptr_t)pvUser;
     unsigned cErrors = 0;
     int rc;
